Added 'dtype=float64' to each empty pd.Series()

+++ pandapipes/test/openmodelica_comparison/pipeflow_openmodelica_comparison.py
@@ -51,12 +51,12 @@ def pipeflow_openmodelica_comparison(net, log_results=True, friction_model='cole
             v_diff_from_pipe, v_diff_to_pipe, v_diff_mean_pipe, v_diff_abs_pipe, \
             v_mean_pandapipes_pipe, v_om_pipe = retrieve_velocity_gas(net, 'pipe')
         else:
-            v_diff_abs_pipe = pd.Series()
-            v_om_pipe = pd.Series()
-            v_mean_pandapipes_pipe = pd.Series()
-            v_diff_from_pipe = pd.Series()
-            v_diff_to_pipe = pd.Series()
-            v_diff_mean_pipe = pd.Series()
+            v_diff_abs_pipe = pd.Series(dtype='float64')
+            v_om_pipe = pd.Series(dtype='float64')
+            v_mean_pandapipes_pipe = pd.Series(dtype='float64')
+            v_diff_from_pipe = pd.Series(dtype='float64')
+            v_diff_to_pipe = pd.Series(dtype='float64')
+            v_diff_mean_pipe = pd.Series(dtype='float64')
 
         diff_results_v_pipe = pd.DataFrame(
             {"diff_v_from_pipe": v_diff_from_pipe, "diff_v_to_pipe": v_diff_to_pipe,
@@ -66,12 +66,12 @@ def pipeflow_openmodelica_comparison(net, log_results=True, friction_model='cole
             v_diff_from_valve, v_diff_to_valve, v_diff_mean_valve, v_diff_abs_valve, \
             v_mean_pandapipes_valve, v_om_valve = retrieve_velocity_gas(net, 'valve')
         else:
-            v_diff_abs_valve = pd.Series()
-            v_om_valve = pd.Series()
-            v_mean_pandapipes_valve = pd.Series()
-            v_diff_from_valve = pd.Series()
-            v_diff_to_valve = pd.Series()
-            v_diff_mean_valve = pd.Series()
+            v_diff_abs_valve = pd.Series(dtype='float64')
+            v_om_valve = pd.Series(dtype='float64')
+            v_mean_pandapipes_valve = pd.Series(dtype='float64')
+            v_diff_from_valve = pd.Series(dtype='float64')
+            v_diff_to_valve = pd.Series(dtype='float64')
+            v_diff_mean_valve = pd.Series(dtype='float64')
 
         diff_results_v_valve = pd.DataFrame(
             {"diff_v_from_valve": v_diff_from_valve, "diff_v_to_valve": v_diff_to_valve,
@@ -85,25 +85,25 @@ def pipeflow_openmodelica_comparison(net, log_results=True, friction_model='cole
                 T_diff_mean_pipe, T_diff_abs_pipe, T_mean_pandapipes_pipe, T_om_pipe = \
                     retrieve_temperature_liquid(net)
         else:
-            v_diff_abs_pipe = pd.Series()
-            v_om_pipe = pd.Series()
-            v_mean_pandapipes_pipe = pd.Series()
-            v_diff_mean_pipe = pd.Series()
+            v_diff_abs_pipe = pd.Series(dtype='float64')
+            v_om_pipe = pd.Series(dtype='float64')
+            v_mean_pandapipes_pipe = pd.Series(dtype='float64')
+            v_diff_mean_pipe = pd.Series(dtype='float64')
 
             if mode != "hydraulics":
-                T_diff_abs_pipe = pd.Series()
-                T_om_pipe = pd.Series()
-                T_mean_pandapipes_pipe = pd.Series()
-                T_diff_mean_pipe = pd.Series()
+                T_diff_abs_pipe = pd.Series(dtype='float64')
+                T_om_pipe = pd.Series(dtype='float64')
+                T_mean_pandapipes_pipe = pd.Series(dtype='float64')
+                T_diff_mean_pipe = pd.Series(dtype='float64')
 
         if 'valve' in net:
             v_diff_mean_valve, v_diff_abs_valve, v_mean_pandapipes_valve, v_om_valve = \
                 retrieve_velocity_liquid(net, element="valve")
         else:
-            v_diff_abs_valve = pd.Series()
-            v_om_valve = pd.Series()
-            v_mean_pandapipes_valve = pd.Series()
-            v_diff_mean_valve = pd.Series()
+            v_diff_abs_valve = pd.Series(dtype='float64')
+            v_om_valve = pd.Series(dtype='float64')
+            v_mean_pandapipes_valve = pd.Series(dtype='float64')
+            v_diff_mean_valve = pd.Series(dtype='float64')
 
         diff_results_v_pipe = pd.DataFrame({"diff_v_mean_pipe": v_diff_mean_pipe,
                                             "diff_v_abs_pipe": v_diff_abs_pipe})
